+++ src/libs/constant/character.tsx
